changement de type

+++ src/interface_type_001_sensors.c
@@ -506,7 +506,7 @@ mea_error_t interface_type_001_sensors_process_xpl_msg(interface_type_001_t *i00
 }
 
 
-int interface_type_001_sensors_poll_inputs(interface_type_001_t *i001, tomysqldb_md_t *md)
+int16_t interface_type_001_sensors_poll_inputs(interface_type_001_t *i001, tomysqldb_md_t *md)
 {
    queue_t *sensors_list=i001->sensors_list;
    struct sensor_s *sensor;
